+++ jest.config.cjs
@@ -5,18 +5,18 @@ const config = require('@zcloak/dev/config/jest.cjs');
 
 module.exports = Object.assign({}, config, {
   moduleNameMapper: {
-    '@zcloak/crypto(.*)$': '<rootDir>/packages/crypto/src/$1',
-    '@zcloak/ctype(.*)$': '<rootDir>/packages/ctype/src/$1',
-    '@zcloak/did-resolver(.*)$': '<rootDir>/packages/did-resolver/src/$1',
-    '@zcloak/did(.*)$': '<rootDir>/packages/did/src/$1',
-    '@zcloak/keyring(.*)$': '<rootDir>/packages/keyring/src/$1',
+    '@zcloak/crypto(.*)$': '<rootDir>/utility/crypto/src/$1',
+    '@zcloak/keyring(.*)$': '<rootDir>/utility/keyring/src/$1',
+    '@zcloak/wasm-asm(.*)$': '<rootDir>/utility/wasm-asm/src/$1',
+    '@zcloak/wasm-bridge(.*)$': '<rootDir>/utility/wasm-bridge/src/$1',
+    '@zcloak/wasm(.*)$': '<rootDir>/utility/wasm/src/$1',
+    '@zcloak/ctype(.*)$': '<rootDir>/protocol/ctype/src/$1',
+    '@zcloak/did-resolver(.*)$': '<rootDir>/protocol/did-resolver/src/$1',
+    '@zcloak/did(.*)$': '<rootDir>/protocol/did/src/$1',
+    '@zcloak/vc(.*)$': '<rootDir>/protocol/vc/src/$1',
     '@zcloak/message(.*)$': '<rootDir>/packages/message/src/$1',
-    '@zcloak/vc(.*)$': '<rootDir>/packages/vc/src/$1',
     '@zcloak/verify(.*)$': '<rootDir>/packages/verify/src/$1',
-    '@zcloak/wasm-asm(.*)$': '<rootDir>/packages/wasm-asm/src/$1',
-    '@zcloak/wasm-bridge(.*)$': '<rootDir>/packages/wasm-bridge/src/$1',
-    '@zcloak/wasm(.*)$': '<rootDir>/packages/wasm/src/$1',
-    'test-support(.*)$': '<rootDir>/packages/test-support/src/$1'
+    'test-support(.*)$': '<rootDir>/test-support/src/$1'
   },
   testTimeout: 30000
 });